requirements.txt
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
enum-compat==0.0.3
Flask==3.1.0
idna==3.10
intuit-oauth==1.2.6
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-quickbooks==0.9.10
rauth==0.7.3
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.7.0
simplejson==3.19.3
six==1.17.0
tailwindcss==0.0.1
urllib3==2.3.0
Werkzeug==3.1.3
